+++ NEWS.md
@@ -48,11 +48,11 @@
 
 * `clean_names()` now supports all object types that have either names or dimnames (#481, @DanChaltiel).
 
-* `make_clean_names()` allows duplicate names to be returned by specifying `unique_sep = NULL` (#495, @JasonAizkalns).
-
 * `adorn_pct_formatting()` uses the locale-dependent value of `decimal.mark` as a decimal separator, e.g., in locales where `getOption("OutDec")` is `,` it will print percentages in the format `"12,34%"` (#451).
 
 * `adorn_totals(where ="row")` now preserves factor class and levels of the first column of the input data.frame (#494). 
+
+* `make_clean_names()` now allows duplicate names to be returned with by specifying `TRUE` to the new `allow_dupes` argument (#495, @JasonAizkalns).
 
 ## Bug fixes